Groundwork for adding image and animation editing functions.
